# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError GenericError is a generic error message returned by a server
    properties:
      message:
        description: message
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/gokusayon/products-api/sdk/models
  Product:
    description: Product Product defines the structure for an API product
    properties:
      description:
        description: the Description of the Product
        maxLength: 10000
        type: string
        x-go-name: Description
      id:
        description: the ID of the product
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: the Name of the Product
        maxLength: 25
        type: string
        x-go-name: Name
      price:
        description: the price for the product
        format: float
        minimum: 0.01
        type: number
        x-go-name: Price
      sku:
        description: the SKU for the product
        pattern: '[a-z]+-[a-z]+-[a-z]+'
        type: string
        x-go-name: SKU
    required:
    - id
    - name
    - price
    - sku
    type: object
    x-go-package: github.com/gokusayon/products-api/sdk/models
  ValidationError:
    description: ValidationError ValidationError is a collection of validation error messages
    properties:
      messages:
        description: messages
        items:
          type: string
        type: array
        x-go-name: Messages
    type: object
    x-go-package: github.com/gokusayon/products-api/sdk/models
info:
  description: Documentation for Product API
  title: for Product API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      parameters:
      - description: |-
          Currency used when returning the price of the product.
          when not specified it returns in GBP
        in: query
        name: currency
        type: string
        x-go-name: Currency
      responses:
        "200":
          $ref: '#/responses/listProducts'
      tags:
      - products
    post:
      description: Adds a product to the list of products
      operationId: addProduct
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes a Product
      operationId: deleteProduct
      parameters:
      - description: The ID of product to be deleted
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    get:
      description: Returns a product from list of products
      operationId: listSingleProduct
      parameters:
      - description: |-
          Currency used when returning the price of the product.
          when not specified it returns in GBP
        in: query
        name: currency
        type: string
        x-go-name: Currency
      - description: The ID of product to be deleted
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    put:
      description: Updates a product
      operationId: updateProduct
      parameters:
      - description: The ID of product to be deleted
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
        "422":
          $ref: '#/responses/errorValidation'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  listProducts:
    description: A list of products returns in response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
  noContent:
    description: ProductsNoContent No content is returned by this API endpoint
  productResponse:
    description: A product returns in response
    schema:
      $ref: '#/definitions/Product'
schemes:
- http
swagger: "2.0"